backend/requirements.txt
fastapi>=0.129.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.31.0
sqladmin>=0.23.0
httpx>=0.26.0
python-multipart>=0.0.9
pydantic-settings>=2.1.0
alembic>=1.13.0
psycopg2-binary
bcrypt==3.2.2
passlib[bcrypt]>=1.7.4
itsdangerous
wtforms
wtforms-sqlalchemy
python-jose[cryptography]
redis>=5.0.0
apscheduler
